tests/requirements.txt
asn1crypto==1.5.1
cached-property==2.0.1
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.3
python-pkcs11==0.7.0
